+++ scripts/workflow/plan.js
@@ -1,0 +1,3 @@
+const { runLifecycleStage } = require('./lifecycle_facade');
+
+runLifecycleStage('plan');